Convert to typescript
