+++ package.json
@@ -38,17 +38,22 @@
     "@angular/platform-browser-dynamic": "~2.0.1",
     "@angular/upgrade": "~2.0.1",
     "angular": "1.5.8",
+    "angular-ui-router": "^1.0.0-beta.3",
     "babel-polyfill": "^6.9.1",
     "babel-runtime": "^6.11.6",
     "core-js": "^2.4.1",
     "oclazyload": "1.0.9",
     "rxjs": "5.0.0-beta.12",
-    "angular-ui-router": "^1.0.0-beta.3",
-    "ui-router-ng2": "^1.0.0-beta.3",
     "ui-router-ng1-to-ng2": "^1.0.12",
+    "ui-router-ng2": "^1.0.0-beta.3",
     "zone.js": "^0.6.25"
   },
   "devDependencies": {
+    "@types/angular": "^1.5.19",
+    "@types/core-js": "^0.9.34",
+    "@types/node": "^6.0.46",
+    "@types/systemjs": "^0.19.31",
+    "awesome-typescript-loader": "^2.2.4",
     "babel-core": "^6.14.0",
     "babel-eslint": "6.1.2",
     "babel-loader": "^6.2.4",
@@ -65,6 +70,7 @@
     "progress-bar-webpack-plugin": "^1.9.0",
     "raw-loader": "^0.5.1",
     "source-map-loader": "^0.1.5",
+    "typescript": "^2.0.9",
     "url-loader": "^0.5.7",
     "webpack": "2.1.0-beta.22",
     "webpack-dev-server": "2.1.0-beta.3",
