--- src/app/common/index.js
@@ -1,5 +1,0 @@
-export * from './common-ng1.module.js';
-export * from './common-ng2.module.js';
-
-export * from './services/admin.service.js';
-export * from './services/login.service.js';
