+++ src/app/common/index.ts
@@ -1,0 +1,5 @@
+export * from './common-ng1.module';
+export * from './common-ng2.module';
+
+export * from './services/admin.service';
+export * from './services/login.service';
