+++ src/app/common/services/admin.service.ts
@@ -1,8 +1,11 @@
 import { Injectable, Inject } from '@angular/core';
+declare var angular: any;
 
 @Injectable()
 export class AdminService {
 
+  $http: any;
+  
   constructor(@Inject('$http') $http) {
     this.$http = $http;
   }
