--- src/app/index.js
@@ -1,3 +1,0 @@
-export * from './app.module.js';
-export * from './upgrade.module.js';
-export * from './app.state.js';
