+++ src/app/upgrade.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { Ng1ToNg2Module } from 'ui-router-ng1-to-ng2';
-import { CommonNg2Module } from 'app/common';
+import { CommonNg2Module } from './common';
 
 @NgModule({
   imports: [
