+++ src/bootstrap.ts
@@ -1,5 +1,5 @@
-import { adapter } from './adapter.js';
-import { appModule } from 'app';
+import { adapter } from './adapter';
+import { appModule } from './app';
 
 document.addEventListener('DOMContentLoaded', () => {
   // `adapter['ng2AppModule']` is `app/upgrade.module.js`
