+++ src/polyfills.ts
@@ -5,4 +5,3 @@ import 'babel-polyfill';
 import 'zone.js/dist/zone';
 import 'zone.js/dist/long-stack-trace-zone';
 
-Error.stackTraceLimit = Infinity;
